refactor: move jwt configuration to a separate file
EPMDJ-10742
